+++ CHANGELOG.rst
@@ -6,7 +6,7 @@ All notable changes to this project will be documented in this file.
 The format is based on `Keep a Changelog <https://keepachangelog.com>`_
 and this project adheres to `Semantic Versioning <https://semver.org>`_.
 
-0.6.0 (2025-09-10)
+0.6.1 (2025-09-10)
 ------------------
 
 * New functions: :class:`~pyunlocbox.functions.proj_lineq`,